Added a new 2D version of multigrid which always has power-of-2 sizes for the local grids, thus avoiding a lot of corner cases.
This version seems to have issues with reproducibility -- the final
residual should be identical but it is not. I need to look for
uninitialized values or something ...

In addition there is a version with a "minimonitoring", that is
recording some events plus timing in order to watch the
scaleup/scaledown/smoothen time depending on the grid level. From this
I'm going to determine the grid level when one wants to switch from Team All
to a smaller team and eventually to one unit for very coarse grids.